refactor(组件): 将应用配置编辑功能抽离为独立组件
将 `AppManageDrawer` 中的应用配置编辑逻辑抽离为独立的 `AppEditDrawer` 组件，以提高代码的可维护性和复用性
